Add new file with all functions deleted
