+++ _str_utils.c
@@ -1,0 +1,107 @@
+#include "shell.h"
+/**
+ *_strcat - copy string to another string.
+ *@dest: char
+ *@src: char
+ *Return: dest
+ *
+ */
+
+char *_strcat(char *dest, char *src)
+{
+	int d = 0, s = 0;
+
+	while (dest[d] != '\0')
+	{
+		d++;
+	}
+	while (src[s] != '\0')
+	{
+		dest[d + s] = src[s];
+		s++;
+	}
+	return (dest);
+}
+
+/**
+ * _strcmp - compare the values of a string
+ * @s1: character
+ * @s2: character
+ * Return: 0
+ */
+
+int _strcmp(char *s1, char *s2)
+{
+	int a;
+
+	for (a = 0; s1[a] != '\0' && s2[a] != '\0'; a++)
+	{
+		if (s1[a] != s2[a])
+			return ((int)s1[a] - s2[a]);
+	}
+	return (0);
+}
+
+/**
+ * _strlen - copies the string pointed to by src into dest
+ * @s: A pointer
+ * Return: char pointer to dest
+ */
+
+int _strlen(char *s)
+{
+	int ch = 0;
+
+	while (*(s + ch) != '\0')
+	{
+		ch++;
+	}
+
+	return (ch);
+}
+
+/**
+ *_strncmp -  function that compares two strings.
+ *@s1: string one
+ *@s2: string two
+ *@n: number of characters
+ * Return: diference
+ */
+
+size_t _strncmp(char *s1, char *s2, size_t n)
+{
+	size_t i, j;
+
+	for (j = 0; s1[j] != '\0' && j < n; j++)
+	{
+		i = s1[j] - s2[j];
+
+		if (i != 0)
+		{
+			return (i);
+		}
+	}
+	return (0);
+}
+
+/**
+ * _strcpy - copies the string pointed to by src into dest
+ * @dest: destination of the copy
+ * @src: source of the copy
+ *
+ * Return: char pointer to dest
+ */
+
+char *_strcpy(char *dest, char *src)
+{
+	int i = 0;
+
+	while (*(src + i) != '\0')
+	{
+		*(dest + i) = *(src + i);
+		++i;
+	}
+	*(dest + i) = *(src + i);
+
+	return (dest);
+}
