--- _strcmp.c
@@ -1,20 +1,0 @@
-#include "shell.h"
-
-/**
- * _strcmp - compare the values of a string
- * @s1: character
- * @s2: character
- * Return: 0
- */
-
-int _strcmp(char *s1, char *s2)
-{
-	int a;
-
-	for (a = 0; s1[a] != '\0' && s2[a] != '\0'; a++)
-	{
-		if (s1[a] != s2[a])
-			return ((int) s1[a] - s2[a]);
-	}
-	return (0);
-}
