--- _strcpy.c
@@ -1,23 +1,0 @@
-#include "shell.h"
-
-/**
- * _strcpy - copies the string pointed to by src into dest
- * @dest: destination of the copy
- * @src: source of the copy
- *
- * Return: char pointer to dest
- */
-
-char *_strcpy(char *dest, char *src)
-{
-	int i = 0;
-
-	while (*(src + i) != '\0')
-	{
-		*(dest + i) = *(src + i);
-		++i;
-	}
-	*(dest + i) = *(src + i);
-
-	return (dest);
-}
