--- _strlen.c
@@ -1,19 +1,0 @@
-#include "shell.h"
-
-/**
- * _strlen - copies the string pointed to by src into dest
- * @s: A pointer
- * Return: char pointer to dest
- */
-
-int _strlen(char *s)
-{
-	int ch = 0;
-
-	while (*(s + ch) != '\0')
-	{
-		ch++;
-	}
-
-	return (ch);
-}
